Adding to github
Done. It is very simple that dismiss any further info. It was made for
learning purposes.

Goo.java and MySimpleGoo.java are the main engine given. The engine to
create a window and print falling forms to it.

TwoDimensionalShapes.java describes a interface to communicate with the
shapes classes. Very simple.

Oval, Rectangle, Circle classes are shape classes that inherits from
the abstract class SimpleShape.